Aggregated 3 reports into one with tabs
