+++ monkey/monkey_island/cc/ui/src/components/report-components/AttackReport.js
@@ -2,36 +2,35 @@ import React from 'react';
 import {Col} from 'react-bootstrap';
 import {ReactiveGraph} from 'components/reactive-graph/ReactiveGraph';
 import {edgeGroupToColor, options} from 'components/map/MapOptions';
-import '../../../styles/Collapse.scss';
-import AuthComponent from '../../AuthComponent';
-import {ScanStatus} from '../../attack/techniques/Helpers';
+import '../../styles/Collapse.scss';
+import AuthComponent from '../AuthComponent';
+import {ScanStatus} from '../attack/techniques/Helpers';
 import Collapse from '@kunukn/react-collapse';
 
-import T1210 from '../../attack/techniques/T1210';
-import T1197 from '../../attack/techniques/T1197';
-import T1110 from '../../attack/techniques/T1110';
-import T1075 from '../../attack/techniques/T1075';
-import T1003 from '../../attack/techniques/T1003';
-import T1059 from '../../attack/techniques/T1059';
-import T1086 from '../../attack/techniques/T1086';
-import T1082 from '../../attack/techniques/T1082';
-import T1145 from '../../attack/techniques/T1145';
-import T1105 from '../../attack/techniques/T1105';
-import T1107 from '../../attack/techniques/T1107';
-import T1065 from '../../attack/techniques/T1065';
-import T1035 from '../../attack/techniques/T1035';
-import T1129 from '../../attack/techniques/T1129';
-import T1106 from '../../attack/techniques/T1106';
-import T1188 from '../../attack/techniques/T1188';
-import T1090 from '../../attack/techniques/T1090';
-import T1041 from '../../attack/techniques/T1041';
-import T1222 from '../../attack/techniques/T1222';
-import T1005 from '../../attack/techniques/T1005';
-import T1018 from '../../attack/techniques/T1018';
-import T1016 from '../../attack/techniques/T1016';
-import T1021 from '../../attack/techniques/T1021';
-import T1064 from '../../attack/techniques/T1064';
-import {extractExecutionStatusFromServerResponse} from '../common/ExecutionStatus';
+import T1210 from '../attack/techniques/T1210';
+import T1197 from '../attack/techniques/T1197';
+import T1110 from '../attack/techniques/T1110';
+import T1075 from '../attack/techniques/T1075';
+import T1003 from '../attack/techniques/T1003';
+import T1059 from '../attack/techniques/T1059';
+import T1086 from '../attack/techniques/T1086';
+import T1082 from '../attack/techniques/T1082';
+import T1145 from '../attack/techniques/T1145';
+import T1105 from '../attack/techniques/T1105';
+import T1107 from '../attack/techniques/T1107';
+import T1065 from '../attack/techniques/T1065';
+import T1035 from '../attack/techniques/T1035';
+import T1129 from '../attack/techniques/T1129';
+import T1106 from '../attack/techniques/T1106';
+import T1188 from '../attack/techniques/T1188';
+import T1090 from '../attack/techniques/T1090';
+import T1041 from '../attack/techniques/T1041';
+import T1222 from '../attack/techniques/T1222';
+import T1005 from '../attack/techniques/T1005';
+import T1018 from '../attack/techniques/T1018';
+import T1016 from '../attack/techniques/T1016';
+import T1021 from '../attack/techniques/T1021';
+import T1064 from '../attack/techniques/T1064';
 
 const tech_components = {
   'T1210': T1210,
@@ -67,35 +66,14 @@ class AttackReportPageComponent extends AuthComponent {
   constructor(props) {
     super(props);
     this.state = {
-      report: false,
-      allMonkeysAreDead: false,
-      runStarted: true,
+      report: this.props.report,
       collapseOpen: ''
     };
   }
 
-  componentDidMount() {
-    this.updateMonkeysRunning().then(res => this.getReportFromServer(res));
-  }
-
-  updateMonkeysRunning = () => {
-    return this.authFetch('/api')
-      .then(res => res.json())
-      .then(res => {
-        this.setState(extractExecutionStatusFromServerResponse(res));
-        return res;
-      });
-  };
-
-  getReportFromServer(res) {
-    if (res['completed_steps']['run_monkey']) {
-      this.authFetch('/api/attack/report')
-        .then(res => res.json())
-        .then(res => {
-          this.setState({
-            report: res
-          });
-        });
+  componentDidUpdate(prevProps) {
+    if (this.props.report !== prevProps.report) {
+     this.setState({ report: this.props.report })
     }
   }
 
@@ -168,31 +146,28 @@ class AttackReportPageComponent extends AuthComponent {
       content.push(this.getTechniqueCollapse(tech_id))
     });
     return (
+      <div id="attack">
+      <h3>
+        ATT&CK report
+      </h3>
+      <p>
+        This report shows information about ATT&CK techniques used by Infection Monkey.
+      </p>
       <div>
         {this.renderLegend()}
         <section className="attack-report">{content}</section>
       </div>
+      <br/>
+    </div>
     )
   }
 
   render() {
-    let content;
-    if (!this.state.runStarted) {
-      content =
-        <p className="alert alert-warning">
-          <i className="glyphicon glyphicon-warning-sign" style={{'marginRight': '5px'}}/>
-          You have to run a monkey before generating a report!
-        </p>;
-    } else if (this.state.report === false) {
-      content = (<h1>Generating Report...</h1>);
-    } else if (Object.keys(this.state.report).length === 0) {
-      if (this.state.runStarted) {
-        content = (<h1>No techniques were scanned</h1>);
-      }
+    if (Object.keys(this.state.report).length === 0 && this.state.runStarted) {
+        return (<h1>No techniques were scanned</h1>);
     } else {
-      content = this.generateReportContent();
+      return (<div> {this.generateReportContent()}</div>);
     }
-    return (<div> {content} </div>);
   }
 }
